chatbot/requirements.txt
# Same as main requirements.txt but organized for the chatbot module

# Core LangChain
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2

# Vector stores
chromadb==0.4.22
faiss-cpu==1.7.4

# Document processing
pypdf==3.17.4
pdfplumber==0.10.3

# OpenAI
openai==1.6.1
tiktoken==0.5.2

# Knowledge graph
networkx==3.2.1

# Utilities
python-dotenv==1.0.0
